doc_parser/requirements.txt
fastapi
uvicorn[standard]

openai
python-dotenv

chromadb
numpy
pandas
openpyxl

matplotlib

pydantic
sqlalchemy

tqdm
